Split tree.py up into its own module
